+++ config/ksql-production-server.properties
@@ -74,9 +74,14 @@ ksql.streams.producer.delivery.timeout.ms=2147483647
 # Kafka cluster is unavailable.
 ksql.streams.producer.max.block.ms=9223372036854775807
 
-# For better fault tolerance and durability, set the replication factor for the KSQL
-# Server's internal topics. Note: the value 3 requires at least 3 brokers in your Kafka cluster.
+# For better fault tolerance and durability, set the replication factor and minimum insync replicas
+# for the KSQL Server's internal topics. All internal topic producers should have acks set to all.
+# All internal topic consumers should
+# Note: the value 3 requires at least 3 brokers in your Kafka cluster.
 ksql.internal.topic.replicas=3
+ksql.internal.topic.min.insync.replicas=2
+ksql.server.command.producer.acks=all
+ksql.server.command.consumer.auto.offset.reset=never
 
 # Configure underlying Kafka Streams internal topics in order to achieve better fault tolerance and
 # durability, even in the face of Kafka broker failures. Highly recommended for mission critical applications.
